patterns: more uniform alternative
remove SMTPatT and use a single pattern constuctor for any "kind". We
cannot make it into an inductive because of universe restrictions, so
rename the marke to `smt_pat` and propagate across the entire library.

This is intentionally a big commit, splitting it would break
bisectability: an F* binary without the new pattern names will not work
with the new ulib (I think specifically because of the invertOption
lemma in prims) and vice versa.